requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
idna==2.9
multidict==4.7.5
oauthlib==3.1.0
olefile==0.46
Pillow==7.1.2
pycparser==2.20
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2020.1
requests>=2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tornado==6.0.4
urllib3>=1.25.9
websockets==8.1
yarl==1.4.2
youtube-dl==2017.3.20
arrow==0.15.6
